requirements.txt
pyufbr
flask
flask-script
flask-sqlalchemy
flask-wtf
flask-login
Flask-Uploads
Flask-Migrate
wtforms-sqlalchemy
Flask-Mail
Flask-Datepicker
flask-bootstrap4